refactor: implement access control for clinic and platform basic users, removing legacy staff checks
