+++ src/collections/PlattformStaff/index.ts
@@ -1,5 +1,4 @@
 import type { CollectionConfig } from 'payload'
-import { isPlatformStaff, isPlatformStaffOrSelf } from '@/access/isStaff'
 
 // Profile collection for Platform Staff members
 export const PlattformStaff: CollectionConfig = {
